Added tableEntity Numbers. Implemented Argument Searching
tableEntity class now incorporates designated table numbers, instead of wrapping them up in the table title as a whole. The constructor also modified to account for this new member variable.

This inclusion of numbers also enabled me to implement argument-based searching as detailed in the instructions. Table numbers can now be searched and matched to tables in tableVector.